modify: split RegisterForm
